+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v1.1.1
  * The Roster is now refreshed on each connection
  * We get the server Caps during the connection
  * Add anonymous login support
+ * Fix an issue during Pubsub subscription
 
 v1.1.0
 ---------------------------